+++ PluginFactory.pro
@@ -13,11 +13,7 @@ CONFIG      += staticlib
 CONFIG      += c++11
 
 HEADERS += \
-    pfactory/error_policy/pfactory_exception.h \
-    pfactory/error_policy/pfactory_ignore_error_policy.h \
-    pfactory/error_policy/pfactory_throw_exception_error_policy.h \
     pfactory/pfactory.h \
-    pfactory/plugin/iobject_creator.h \
     pfactory/plugin/object_creator.h \
     pfactory/plugin/psys/iplugin.h \
     pfactory/plugin/psys/isub_plugin.h \
@@ -25,7 +21,6 @@ HEADERS += \
     pfactory/plugin/psys/loader/plugin_loader.h \
     pfactory/plugin/psys/loader/plugin_loader_private.h \
     pfactory/plugin/psys/plugin_data.h \
-    pfactory/plugin/psys/sub_plugin_info.h \
     pfactory/plugin/sub_plugin_creator.h \
     pfactory/plugin/template_plugin.h \
     pfactory/plugin_interfaces/icreator_factory_plugin_loader.h \
@@ -33,21 +28,18 @@ HEADERS += \
     pfactory/private_factory/internal_object_creator_private.h \
     pfactory/private_factory/pfactory_interface_private.h \
     pfactory/private_factory/pfactory_plugin_controller_private.h \
-    pfactory/private_factory/pfactory_private.h \
-    pfactory/private_factory/pfactory_singleton.h \
+    pfactory/private_factory/pfactory_policy.h \
     pfactory/private_factory/pfactory_template_interface_private.h \
     pfactory/private_factory/pfactory_template_plug_creator_private.h
 
 
 SOURCES += \
-    pfactory/error_policy/pfactory_exception.cpp \
     pfactory/plugin/psys/iplugin.cpp \
     pfactory/plugin/psys/isub_plugin.cpp \
     pfactory/plugin/psys/loader/iplugin_loader.cpp \
     pfactory/plugin/psys/loader/plugin_loader.cpp \
     pfactory/plugin/psys/loader/plugin_loader_private.cpp \
     pfactory/plugin/psys/plugin_data.cpp \
-    pfactory/plugin/psys/sub_plugin_info.cpp \
     pfactory/plugin/sub_plugin_creator.cpp \
     pfactory/plugin/template_plugin.cpp \
     pfactory/plugin_interfaces/icreator_factory_plugin_loader.cpp \
