--- pfactory/error_policy/pfactory_exception.cpp
@@ -1,24 +1,0 @@
-#include "pfactory_exception.h"
-using namespace pf;
-
-PFactoryException::PFactoryException(const std::string &msg) noexcept
-  : _msg( msg )
-{
-
-}
-
-PFactoryException::PFactoryException(const QString &msg) noexcept
-  : _msg( msg.toStdString () )
-{
-
-}
-
-PFactoryException::~PFactoryException() noexcept
-{
-
-}
-
-const char *PFactoryException::what() const noexcept
-{
-  return _msg.c_str();
-}
