+++ pfactory/plugin/psys/isub_plugin.h
@@ -1,12 +1,44 @@
 #pragma once
+#include <memory>
 #include <QString>
 #include <QMetaClassInfo>
-#include <memory>
-
-#include "sub_plugin_info.h"
 
 namespace psys
 {
+  ///
+  /// \brief Information about the sub plugin
+  ///
+  struct SubPluginInfo final
+  {
+    ////
+    /// \brief Sub plugin id
+    ///
+    QString id;
+
+    ///
+    /// \brief ID of the implementing interface
+    ///
+    QString interfaceId;
+
+    // constructs
+    SubPluginInfo();
+    SubPluginInfo(const QString &id,
+                  const QString &interface);
+    SubPluginInfo(const SubPluginInfo &info);
+
+    // methods
+    bool isValid() const;
+
+    // Operators
+    bool operator==(const QString &id) const;
+    bool operator!=(const QString &id) const;
+
+    bool operator==(const SubPluginInfo &info) const;
+    bool operator!=(const SubPluginInfo &info) const;
+
+    bool operator>(const SubPluginInfo &info) const;
+    bool operator<(const SubPluginInfo &info) const;
+  };
 
   class ISubPlugin;
   using SubPlugin = std::shared_ptr<ISubPlugin>;
@@ -21,7 +53,7 @@ namespace psys
     ///
     SubPluginInfo   _subPluginInfo;
   protected:
-    ISubPlugin( SubPluginInfo info );
+    ISubPlugin(SubPluginInfo info);
   public:
     virtual ~ISubPlugin();
   public:
